pyproject.toml
[project]
name = "qwed-new"
version = "0.1.0"
description = "Enterprise-grade deterministic verification engine for AI."
authors = [
    {name = "QWED Team", email = "dev@qwed.tech"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic[email]>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.1.0",
    "python-dotenv>=1.0.0",
    "sympy>=1.12",
    "z3-solver>=4.12.0",
    "networkx>=3.0",
    "sqlmodel>=0.0.14",
    "bcrypt>=4.0.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.6",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "psycopg2-binary>=2.9.9",
    "sqlglot>=20.0.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "matplotlib>=3.7.0",
    "docker>=7.0.0",
    "locust>=2.24.0",
    "pytest-timeout>=4.0.0",
    "pytest-xdist>=3.5.0",
    "jinja2>=3.1.3",
    "freezegun>=1.4.0"
]
requires-python = ">=3.10"
readme = "README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qwed_new"]
